requirements.txt
azure-functions==1.19.0
azure-storage-blob==12.20.0
azure-cosmos==4.7.0
pyodbc>=5.3.0
SQLAlchemy==2.0.36
cryptography==43.0.3
python-dotenv==1.0.1
opencensus-ext-azure==1.1.11

